mastodon-streamer/requirements.txt
Mastodon.py
kafka-python
